backend/requirements.txt
markdownify
smolagents
smolagents[gradio]
requests
duckduckgo_search
pandas

PyPDF2

pymsaviz
streamlit

langchain
langchain-community
langchain-openai
langgraph

seaborn

biopython
tabulate

pytest
pytest-asyncio
python-dotenv

# MCP (Model Context Protocol) dependencies
mcp>=1.0.0
fastapi
uvicorn